[giveaways]delete old folder
