--- gaw/__init__.py
@@ -1,5 +1,0 @@
-from .gset import Gset
-
-async def setup(bot):
-    cog = await Gset.initialize(bot)
-    bot.add_cog(cog)
