+++ giveaways/__init__.py
@@ -1,11 +1,5 @@
-import json
-from pathlib import Path
-
-from .giveaway import giveaways
-
-with open(Path(__file__).parent / "info.json") as fp:
-    __red_end_user_data_statement__ = json.load(fp)["end_user_data_statement"]
-
+from .gset import Gset
 
 async def setup(bot):
-    bot.add_cog(await giveaways.inititalze(bot))
+    cog = await Gset.initialize(bot)
+    bot.add_cog(cog)
