+++ giveaways/constants.py
