--- giveaways/info.json
@@ -1,22 +1,0 @@
-
-{
-    "name": "Giveaways",
-    "short": "Start reaction based giveaways.",
-    "description": "Reaction based giveaways with customisable embeds, reactions, amari bot requirements and much more. ",
-    "end_user_data_statement": "This cog stores user ids to link them to their hosted giveaways.",
-    "install_msg": "Thanks for installing Giveaways. Load the cog with `[p]load giveaways` and see all its commands with `[p]help Giveaways`",
-    "author": [
-        "crayyy_zee"
-    ],
-    "required_cogs": {},
-    "requirements": ["git+https://github.com/i-am-zaidali/AmariWrapper.git", "dateparser"],
-    "tags": [
-        "giveaways",
-        "requirements",
-        "raffles",
-        "raffle",
-        "giveaway",
-        "reactions"
-    ],
-    "type": "COG"
-}
