+++ giveaways/main.py
