+++ giveaways/models/guildsettings.py
