+++ giveaways/models/requirements.py
