+++ giveaways/utils.py
